+++ css/main.scss
@@ -90,3 +90,13 @@
         }
     }
 }
+
+#playDiv {
+
+    .player {
+        display: inline-block;
+        width: 250px;
+        background-color: rgb(167, 220, 241);
+        padding: 5px;
+    }
+}
